Feat : 알림리스트 섹션 컴포넌트 생성 및 해상도별 분기처리 (api호출 한번만 할 수 있도록)
