+++ stores/AlertStore.ts
@@ -2,6 +2,7 @@ import { create } from 'zustand';
 
 interface AlertStore {
   isBellDisabled: boolean;
+  alertCount: number;
   handleNewAlert: (alert: string) => void;
   handleMarkAllAsRead: () => void;
   setBellDisabled: (disabled: boolean) => void;
@@ -9,6 +10,10 @@ interface AlertStore {
 
 export const useAlertStore = create<AlertStore>((set) => ({
   isBellDisabled: false,
+  alertCount: 0,
+  setAlertCount: (alertCount: number) => {
+    set({ alertCount: alertCount });
+  },
   handleNewAlert: (alert: string) => {
     set({ isBellDisabled: false });
   },
